[BE/Fix] 조명 색상 hue 값 float로 변경

See merge request s12-final/S12P31D103!121

+++ BE/device-service/src/main/java/com/hogumiwarts/lumos/device/dto/light/LightDetailResponse.java
@@ -42,7 +42,7 @@ public class LightDetailResponse {
     private Integer lightTemperature;
 
     @Schema(description = "조명 색조")
-    private Integer hue;
+    private Float hue;
 
     @Schema(description = "조명 채도")
     private Float saturation;
